pyproject.toml
[project]
name = "barcode-generator"
version = "0.1.0"
description = "A Python program to generate barcode images with labels"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-barcode>=0.15.1",
    "Pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.8",
    "black>=24.8.0",
    "isort>=5.13.2",
]

[tool.black]
line-length = 120
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "F",  # Pyflakes
    "E",  # Pycodestyle (errors)
    "W",  # Pycodestyle (warnings)
    "C90",  # McCabe complexity
]
mccabe.max-complexity = 7

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
